# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserProfile": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"phone"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProfileRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UserSettingsModelType": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"settingsValue": {
						"type": "string"
					},
					"settingsKey": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"settingsValue",
					"settingsKey"
				],
				"type": "object",
				"description": "User Settings"
			},
			"CreateSettingRequest": {
				"properties": {
					"key": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"key",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateSettingRequest": {
				"properties": {
					"value": {
						"type": "string"
					}
				},
				"required": [
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserAddressModelType": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"addressContactPhone": {
						"type": "string"
					},
					"addressContactName": {
						"type": "string"
					},
					"addressLandmark": {
						"type": "string"
					},
					"addressAddressType": {
						"type": "string"
					},
					"addressLongitude": {
						"type": "number",
						"format": "double"
					},
					"addressLatitude": {
						"type": "number",
						"format": "double"
					},
					"addressPostalCode": {
						"type": "number",
						"format": "double"
					},
					"addressCountry": {
						"type": "string"
					},
					"addressState": {
						"type": "string"
					},
					"addressCity": {
						"type": "string"
					},
					"addressStreet": {
						"type": "string"
					},
					"addressLine2": {
						"type": "string"
					},
					"addressLine1": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"addressContactPhone",
					"addressContactName",
					"addressLandmark",
					"addressAddressType",
					"addressLongitude",
					"addressLatitude",
					"addressPostalCode",
					"addressCountry",
					"addressState",
					"addressCity",
					"addressStreet",
					"addressLine2",
					"addressLine1"
				],
				"type": "object",
				"description": "User Address"
			},
			"CreatedAddressResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"address": {
						"$ref": "#/components/schemas/UserAddressModelType"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserAddressRequest": {
				"properties": {
					"address_line1": {
						"type": "string"
					},
					"address_line2": {
						"type": "string"
					},
					"street": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"latitude": {
						"type": "string"
					},
					"longitude": {
						"type": "string"
					},
					"addressType": {
						"type": "string"
					},
					"landmark": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					}
				},
				"required": [
					"address_line1",
					"city",
					"state",
					"country",
					"latitude",
					"longitude",
					"postalCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SearchRequest": {
				"properties": {
					"query": {
						"type": "string"
					},
					"aiStoreId": {
						"type": "string"
					}
				},
				"required": [
					"query",
					"aiStoreId"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "core-api",
		"version": "1.4.0",
		"description": "API documentation for the KiranaPro - 'PROD' Backend microservice, designed for seamless integration with multiple microservices, including AI, ONDC/ KIKO/ third-party ONDC services, and mobile frontends (iOS & Android).",
		"license": {
			"name": "Proprietary - Copyright (c) 2024, KiranaPro"
		},
		"contact": {
			"name": "Sandeep K ",
			"email": "sandeep@kirana.pro"
		}
	},
	"paths": {
		"/user": {
			"post": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "The updated profile.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedProfile": {
											"$ref": "#/components/schemas/UserProfile"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"updatedProfile",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Invalid inputs"
					}
				},
				"description": "Update the user's profile settings.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The updated profile data (name and/or email).",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProfileRequest",
								"description": "The updated profile data (name and/or email)."
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "The user's profile.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserProfile"
								}
							}
						}
					}
				},
				"description": "Get the user's profile settings.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/user/settings": {
			"get": {
				"operationId": "GetUserSettings",
				"responses": {
					"200": {
						"description": "An array of user settings.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserSettingsModelType"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieve all user settings.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUserSetting",
				"responses": {
					"200": {
						"description": "A success message.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Invalid inputs"
					}
				},
				"description": "Create a new user setting.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The key and value of the setting.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSettingRequest",
								"description": "The key and value of the setting."
							}
						}
					}
				}
			}
		},
		"/user/settings/{key}": {
			"put": {
				"operationId": "UpdateUserSetting",
				"responses": {
					"200": {
						"description": "A success message or an error if the setting is not found.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Setting not found"
					}
				},
				"description": "Update an existing user setting.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The key of the setting to update.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The updated value.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSettingRequest",
								"description": "The updated value."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUserSetting",
				"responses": {
					"200": {
						"description": "A success message or an error if the setting is not found.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Setting not found"
					}
				},
				"description": "Delete a user setting.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The key of the setting to delete.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/address": {
			"get": {
				"operationId": "GetUserAddresses",
				"responses": {
					"200": {
						"description": "An array of user addresses.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserAddressModelType"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all user addresses.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUserAddress",
				"responses": {
					"200": {
						"description": "A success message and the created address.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreatedAddressResponse"
								}
							}
						}
					},
					"201": {
						"description": "Missing or invalid inputs"
					},
					"400": {
						"description": "Missing or invalid inputs"
					}
				},
				"description": "Create a new user address.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The address details.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserAddressRequest",
								"description": "The address details."
							}
						}
					}
				}
			}
		},
		"/user/address/{id}": {
			"get": {
				"operationId": "GetUserAddressById",
				"responses": {
					"200": {
						"description": "The address or an error if not found.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"address": {
											"$ref": "#/components/schemas/UserAddressModelType"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Address not found"
					}
				},
				"description": "Get a specific user address by ID.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The address ID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUserAddress",
				"responses": {
					"200": {
						"description": "A success message or an error if the address is not found.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"address": {
											"$ref": "#/components/schemas/UserAddressModelType"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Address not found"
					}
				},
				"description": "Delete a user address by ID.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The address ID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/ai/{taskId}": {
			"get": {
				"operationId": "GetCartStatus",
				"responses": {
					"200": {
						"description": "The cart status or cart details.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "Task ID missing!"
					}
				},
				"description": "Fetches the cart status or details based on task ID.",
				"tags": [
					"AI"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The unique task ID.",
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/ai/search": {
			"post": {
				"operationId": "SearchItems",
				"responses": {
					"200": {
						"description": "The search results.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "Missing or invalid inputs!"
					}
				},
				"description": "Searches for items based on a query and AI store ID.",
				"tags": [
					"AI"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The search request containing the query and AI store ID.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SearchRequest",
								"description": "The search request containing the query and AI store ID."
							}
						}
					}
				}
			}
		},
		"/ai": {
			"get": {
				"operationId": "GetPincodeAvailability",
				"responses": {
					"200": {
						"description": "Whether stores are available in the pincode.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"pincode": {
											"type": "boolean"
										}
									},
									"required": [
										"pincode"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Missing or invalid User-Agent header."
					}
				},
				"description": "Retrieves the pincode from the User-Agent header.",
				"tags": [
					"AI"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The User-Agent header.",
						"in": "query",
						"name": "userAgent",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/v1"
		}
	]
}